+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.1.7
 
+Add Toggle builder
+
 ## 0.1.6
 
 * Add support of UIKitButton size and group