+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.1.7
+
+Add Toggle builder
+
 ## 0.1.6
 
 * Add support of UIKitButton size and group